Improvement: Make hide fire particles and hide blaze particles work on 1.21
