+++ src/main/java/at/hannibal2/skyhanni/mixins/transformers/MixinBlockFire.java
@@ -12,7 +12,7 @@ import org.spongepowered.asm.mixin.injection.Redirect;
 public class MixinBlockFire {
 
     @Redirect(method = "randomDisplayTick", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/World;spawnParticle(Lnet/minecraft/util/EnumParticleTypes;DDDDDD[I)V"))
-    private void onLivingUpdate(World world, EnumParticleTypes particleType, double x, double y, double z, double xOffset, double yOffset, double zOffset, int[] parameters) {
+    private void onRandomDisplayTick(World world, EnumParticleTypes particleType, double x, double y, double z, double xOffset, double yOffset, double zOffset, int[] parameters) {
         if (!SkyHanniMod.feature.misc.particleHiders.hideFireBlockParticles) {
             world.spawnParticle(particleType, x, y, z, xOffset, yOffset, zOffset, parameters);
         }
